Moved the masterplan files
I moved the master plan file to our repo. so that we can update it together with the codebase,

notes :
I thnk if we will put some images there, we should also put it inside the folder.
because you can literally copy an image and paste it inside proj, i dont know if it will work tho.